backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Pillow==10.4.0
django-allauth
requests
PyJWT
cryptography
djangorestframework-simplejwt[blacklist]
dj-rest-auth[social]>=2.2.0
stripe>=8.0.0
# Firebase para notificaciones push
firebase-admin>=6.0.0
# IA para búsqueda semántica
sentence-transformers>=2.2.0
numpy>=1.24.0
# Machine Learning para dashboard predictivo
scikit-learn>=1.3.0
pandas>=2.1.0
joblib>=1.3.0
# Generación de reportes
reportlab>=4.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
factory-boy>=3.3.0
# Azure Computer Vision para búsqueda por imagen
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.7.1